+++ model/model.h
@@ -20,10 +20,10 @@ public:
     const Vec3f& GetVertex(const int index) const{return vertices_[index];}    
     const Vec3f& GetNormal(const int index) const{return normals_[index];}    
     const Vec2f& GetTexture(const int index) const{return textures_[index];}    
-    const vector<Vec3f>& GetSurfece(const int index) const{return surfaces_[index];};
+    const vector<Vec3i>& GetSurfece(const int index) const{return surfaces_[index];};
 
 private:
-    vector<vector<Vec3f>> surfaces_;
+    vector<vector<Vec3i>> surfaces_;
     vector<Vec3f> vertices_;
     vector<Vec3f> normals_;
     vector<Vec2f> textures_;
